requirements.txt
pywal
requests
urllib3
pillow